create registration request directive and implement with rejected registration requests
